pyproject.toml
[project]
name = "qna-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.2",
    "fastapi>=0.127.0",
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.0",
    "langsmith==0.4.55",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.5.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "qdrant-client>=1.16.2",
    "ragas==0.3.7",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.40.0",
]
